requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.0
flask>=2.0.0
gradio>=3.0.0
accelerate>=0.20.0
datasets>=2.0.0